# api-docs/my-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "植物种植助手小程序后端API文档",
    "description": "植物种植助手小程序后端服务接口文档，基于Spring Boot 2.7.18实现。",
    "contact": {
      "name": "zszleon",
      "email": "2079973876@qq.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "TokenAuth": []
    }
  ],
...
  "paths": {
...
    "/api/plants/update": {
      "post": {
        "tags": [
          "植物管理"
        ],
        "summary": "更新植物",
        "description": "更新植物信息，需要Token认证",
        "operationId": "updatePlant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePlant"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/plants/page": {
      "post": {
        "tags": [
          "植物管理"
        ],
        "summary": "获取植物列表",
        "description": "分页获取当前用户的植物列表，需要Token认证",
        "operationId": "getPlantsPage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponsePlant"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/plants/getPlantById": {
      "post": {
        "tags": [
          "植物管理"
        ],
        "summary": "获取植物详情",
        "description": "根据植物ID获取植物详细信息，需要Token认证",
        "operationId": "getPlantById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Id"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePlant"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/plants/delete": {
      "post": {
        "tags": [
          "植物管理"
        ],
        "summary": "删除植物",
        "description": "删除指定植物，需要Token认证",
        "operationId": "deletePlant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Id"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/plants/createPlant": {
      "post": {
        "tags": [
          "植物管理"
        ],
        "summary": "创建植物",
        "description": "创建新的植物信息，需要Token认证",
        "operationId": "createPlant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePlant"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "统一API响应"
      },
...
      "Id": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "业务ID",
            "format": "int64",
            "example": 123456789
          }
        },
        "description": "ID请求参数"
      },
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "string",
            "description": "响应数据"
          }
        },
        "description": "统一API响应"
      },
...
      "PlantRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "植物ID，更新时必填",
            "format": "int64",
            "example": 123456789
          },
          "name": {
            "type": "string",
            "description": "植物名称",
            "example": "多肉植物"
          },
          "type": {
            "type": "string",
            "description": "植物类型",
            "example": "多肉植物"
          },
          "description": {
            "type": "string",
            "description": "植物描述",
            "example": "这是一盆可爱的多肉植物"
          },
          "imageUrl": {
            "type": "string",
            "description": "植物图片URL",
            "example": "https://example.com/image.jpg"
          },
          "plantDate": {
            "type": "string",
            "description": "种植日期",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "description": "植物状态",
            "example": "healthy",
            "enum": [
              "healthy",
              "growing",
              "need-care",
              "diseased"
            ]
          },
          "wateringInterval": {
            "type": "integer",
            "description": "浇水间隔天数",
            "format": "int32",
            "example": 7
          },
          "fertilizingInterval": {
            "type": "integer",
            "description": "施肥间隔天数",
            "format": "int32",
            "example": 30
          },
          "lastWatering": {
            "type": "string",
            "description": "上次浇水时间",
            "format": "date-time"
          },
          "lastFertilizing": {
            "type": "string",
            "description": "上次施肥时间",
            "format": "date-time"
          }
        },
        "description": "植物信息请求DTO"
      },
      "ApiResponsePlant": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/Plant"
          }
        },
        "description": "统一API响应"
      },
      "Plant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "plantDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "wateringInterval": {
            "type": "integer",
            "format": "int32"
          },
          "fertilizingInterval": {
            "type": "integer",
            "format": "int32"
          },
          "lastWatering": {
            "type": "string",
            "format": "date-time"
          },
          "lastFertilizing": {
            "type": "string",
            "format": "date-time"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "响应数据"
      },
      "PageQueryRequest": {
        "type": "object",
        "properties": {
          "pageNum": {
            "minimum": 1,
            "type": "integer",
            "description": "页码",
            "format": "int32",
            "example": 1
          },
          "pageSize": {
            "minimum": 1,
            "type": "integer",
            "description": "每页条数",
            "format": "int32",
            "example": 10
          }
        },
        "description": "分页查询请求基类"
      },
      "ApiResponsePageResponsePlant": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponsePlant"
          }
        },
        "description": "统一API响应"
      },
      "PageResponsePlant": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "description": "数据列表",
            "items": {
              "$ref": "#/components/schemas/Plant"
            }
          },
          "total": {
            "type": "integer",
            "description": "总记录数",
            "format": "int64",
            "example": 100
          },
          "pageNum": {
            "type": "integer",
            "description": "当前页码",
            "format": "int32",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "description": "每页条数",
            "format": "int32",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "description": "总页数",
            "format": "int32",
            "example": 10
          }
        },
        "description": "分页响应"
      },
...
    },
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "description": "用户认证Token，通过微信登录接口获取",
        "name": "X-Token",
        "in": "header"
      }
    }
  }
}