# src/main/java/com/tencent/wxcloudrun/controller/PlantController.java
package com.tencent.wxcloudrun.controller;

import com.tencent.wxcloudrun.auth.LoginUser;
import com.tencent.wxcloudrun.dto.req.CommonRequest;
import com.tencent.wxcloudrun.dto.resp.ApiResponse;
import com.tencent.wxcloudrun.dto.resp.PageResponse;
import com.tencent.wxcloudrun.dto.req.PageQueryRequest;
import com.tencent.wxcloudrun.dto.req.PlantRequest;
import com.tencent.wxcloudrun.dto.resp.PlantResponse;
import com.tencent.wxcloudrun.model.Plant;
import com.tencent.wxcloudrun.service.PlantService;
import io.swagger.v3.oas.annotations.Parameter;
import lombok.extern.slf4j.Slf4j;
import org.springframework.web.bind.annotation.*;

import javax.annotation.Resource;
import javax.validation.Valid;

import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.Operation;

/**
 * 植物管理控制器
 * @author zszleon
 */
@Tag(name = "植物管理", description = "植物的增删改查、分页查询等接口")
@Slf4j
@RestController
@RequestMapping("/api/plants")
public class PlantController {

    @Resource
    private PlantService plantService;

    /**
     * 获取用户的所有植物（分页）
     */
    @Operation(summary = "获取植物列表", description = "分页获取当前用户的植物列表")
    @PostMapping("/page")
    public ApiResponse<PageResponse<PlantResponse>> getPlantsPage(@Valid @RequestBody PageQueryRequest request,
                                                                  @Parameter(hidden = true) LoginUser loginUser) {
        log.info("获取植物列表, 用户: {}, 页码: {}, 每页大小: {}", loginUser.getUserId(), request.getPageNum(), request.getPageSize());
        request.setUserId(loginUser.getUserId());
        PageResponse<PlantResponse> pageData = plantService.getPlantsByUserId(request);
        log.info("{}",pageData);
        return ApiResponse.pageOk(pageData);
    }

    /**
     * 根据ID获取植物详情
     */
    @Operation(summary = "获取植物详情", description = "根据植物ID获取植物详细信息")
    @PostMapping("/getPlantById")
    public ApiResponse<PlantResponse> getPlantById(@RequestBody CommonRequest.Id id,
                                           @Parameter(hidden = true) LoginUser loginUser) {
        log.info("获取植物详情, 用户: {}, 植物ID: {}", loginUser.getUserId(), id.getId());
        PlantResponse plant = plantService.getPlantById(id.getId(), loginUser.getUserId());
        if (plant != null) {
            return ApiResponse.ok(plant);
        } else {
            return ApiResponse.error("植物不存在或无权访问");
        }
    }

    /**
     * 创建新植物
     */
    @Operation(summary = "创建植物", description = "创建新的植物信息")
    @PostMapping("/createPlant")
    public ApiResponse<PlantResponse> createPlant(@RequestBody PlantRequest request,
                                          @Parameter(hidden = true) LoginUser loginUser) {
        // 设置用户ID
        request.setUserId(loginUser.getUserId());
        log.info("创建植物, 用户: {}, 植物名称: {}", loginUser.getUserId(), request.getName());
        PlantResponse plant = plantService.createPlant(request);
        return ApiResponse.ok(plant);
    }

    /**
     * 更新植物信息
     */
    @Operation(summary = "更新植物", description = "更新植物信息")
    @PostMapping("/update")
    public ApiResponse<PlantResponse> updatePlant(@RequestBody PlantRequest request,
                                          @Parameter(hidden = true) LoginUser loginUser) {
        // 设置用户ID
        request.setUserId(loginUser.getUserId());
        log.info("更新植物, 用户: {}, 植物ID: {}, 植物名称: {}", loginUser.getUserId(), request.getId(), request.getName());
        PlantResponse plant = plantService.updatePlant(request.getId(), request);
        if (plant != null) {
            return ApiResponse.ok(plant);
        } else {
            return ApiResponse.error("植物不存在或无权修改");
        }
    }

    /**
     * 删除植物
     */
    @Operation(summary = "删除植物", description = "删除指定植物")
    @PostMapping("/delete")
    public ApiResponse<String> deletePlant(@RequestBody CommonRequest.Id id,
                                           @Parameter(hidden = true) LoginUser loginUser) {
        log.info("删除植物, 用户: {}, 植物ID: {}", loginUser.getUserId(), id.getId());
        boolean success = plantService.deletePlant(id.getId(), loginUser.getUserId());
        if (success) {
            return ApiResponse.ok("删除成功");
        } else {
            return ApiResponse.error("植物不存在或无权删除");
        }
    }
}